requirements.txt
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
fastapi==0.104.0
fastapi-utils==0.2.1
greenlet==3.0.0
h11==0.14.0
idna==3.4
pydantic==1.10.13
pydantic_core==2.10.1
python-dotenv==1.0.0
python-multipart==0.0.6
sniffio==1.3.0
SQLAlchemy==1.4.49
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2
